Refactor spell related code
