+++ core/src/main/kotlin/no/ntnu/beardblaster/screen/GameplayScreen.kt
@@ -31,9 +31,7 @@ import no.ntnu.beardblaster.commons.State
 import no.ntnu.beardblaster.game.GameData
 import no.ntnu.beardblaster.game.GameInstance
 import no.ntnu.beardblaster.game.GamePhase
-import no.ntnu.beardblaster.hud.SpellBar
-import no.ntnu.beardblaster.hud.spellbar
-import no.ntnu.beardblaster.spell.SpellLockState
+import no.ntnu.beardblaster.spell.*
 import no.ntnu.beardblaster.sprites.WizardTexture
 import no.ntnu.beardblaster.sprites.WizardTextures
 import no.ntnu.beardblaster.ui.*
