+++ core/src/main/kotlin/no/ntnu/beardblaster/spell/SpellInfo.kt
@@ -1,4 +1,4 @@
-package no.ntnu.beardblaster.ui
+package no.ntnu.beardblaster.spell
 
 import com.badlogic.gdx.scenes.scene2d.ui.Label
 import com.badlogic.gdx.scenes.scene2d.ui.Table
@@ -11,9 +11,7 @@ import no.ntnu.beardblaster.WORLD_HEIGHT
 import no.ntnu.beardblaster.WORLD_WIDTH
 import no.ntnu.beardblaster.assets.Nls
 import no.ntnu.beardblaster.commons.spell.Spell
-import no.ntnu.beardblaster.hud.ElementChangedObserver
-import no.ntnu.beardblaster.models.SpellCasting
-import no.ntnu.beardblaster.spell.SpellLockState
+import no.ntnu.beardblaster.ui.*
 
 private val log = logger<SpellInfo>()
